PA4: working

+++ pa4/src/miniJava/CodeGenerator/RuntimeEntity.java
@@ -1,5 +1,5 @@
 package miniJava.CodeGenerator;
 
 public abstract class RuntimeEntity { 
-    public short size;
+    public int size;
 }
